requirements.txt
streamlit
concrete-python
